+++ R/loo.R
@@ -75,7 +75,7 @@
 #'   distributional assumption and integrate over uncertainty in the parameters.
 #'   This only assumes that any one observation can be omitted without having a
 #'   major effect on the posterior distribution, which can be judged using the
-#'   diagnostic plot provided by the \code{\link[loo]{plot.loo}} method and the
+#'   diagnostic plot provided by the \code{\link[loo:pareto-k-diagnostic]{plot.loo}} method and the
 #'   warnings provided by the \code{\link[loo]{print.loo}} method (see the
 #'   \emph{How to Use the rstanarm Package} vignette for an example of this
 #'   process).
@@ -120,6 +120,7 @@
 #' }
 #'
 #' @examples
+#' if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
 #' \donttest{
 #' fit1 <- stan_glm(mpg ~ wt, data = mtcars, refresh = 0)
 #' fit2 <- stan_glm(mpg ~ wt + cyl, data = mtcars, refresh = 0)
@@ -176,7 +177,7 @@
 #' loo_list <- list(A = loo1, B = loo2, C = loo3) # names optional (affects printing)
 #' loo_model_weights(loo_list)
 #' }
-#'
+#' }
 loo.stanreg <-
   function(x,
            ...,
@@ -640,7 +641,7 @@ reloo <- function(x, loo_x, obs, ..., refit = TRUE) {
         newdata = d[omitted, , drop = FALSE],
         offset = x$offset[omitted],
         newx = get_x(x)[omitted, , drop = FALSE],
-        newz = x$z[omitted, , drop = FALSE], # NULL other than for some stan_betareg models
+        newz = x[["z"]][omitted, , drop = FALSE], # NULL other than for some stan_betareg models
         stanmat = as.matrix.stanreg(fit_j)
       )
   }
@@ -708,6 +709,10 @@ kfold_and_reloo_data <- function(x) {
   } else {
     # already a data frame
     all_vars <- all.vars(formula(x))
+    if (isTRUE(x$stan_function == "stan_gamm4")) {
+      # see https://github.com/stan-dev/rstanarm/issues/435
+      all_vars <- c(all_vars, all.vars(getCall(x)[["random"]]))
+    }
     if ("." %in% all_vars) {
       all_vars <- seq_len(ncol(d))
     }