+++ R/plots.R
@@ -18,9 +18,9 @@
 #' Plot method for stanreg objects
 #'
 #' The \code{plot} method for \link{stanreg-objects} provides a convenient 
-#' interface to the \link[bayesplot]{MCMC} module in the \pkg{\link{bayesplot}} 
-#' package for plotting MCMC draws and diagnostics. It is also straightforward 
-#' to use the functions from the \pkg{bayesplot} package directly rather than 
+#' interface to the \link[bayesplot:MCMC-overview]{MCMC} module in the 
+#' \pkg{\link{bayesplot}} package for plotting MCMC draws and diagnostics. It is also 
+#' straightforward to use the functions from the \pkg{bayesplot} package directly rather than
 #' via the \code{plot} method. Examples of both methods of plotting are given
 #' below.
 #'
@@ -31,18 +31,18 @@
 #' @template args-pars
 #' @template args-regex-pars
 #' @param plotfun A character string naming the \pkg{bayesplot} 
-#'   \link[bayesplot]{MCMC} function to use. The default is to call
-#'   \code{\link[bayesplot]{mcmc_intervals}}. \code{plotfun} can be specified
+#'   \link[bayesplot:MCMC-overview]{MCMC} function to use. The default is to call
+#'   \code{\link[bayesplot:MCMC-intervals]{mcmc_intervals}}. \code{plotfun} can be specified
 #'   either as the full name of a \pkg{bayesplot} plotting function (e.g.
 #'   \code{"mcmc_hist"}) or can be abbreviated to the part of the name following
 #'   the \code{"mcmc_"} prefix (e.g. \code{"hist"}). To get the names of all
-#'   available MCMC functions see \code{\link[bayesplot]{available_mcmc}}.
+#'   available MCMC functions see \code{\link[bayesplot:available_ppc]{available_mcmc}}.
 #'
 #' @param ... Additional arguments to pass to \code{plotfun} for customizing the
 #'   plot. These are described on the help pages for the individual plotting 
 #'   functions. For example, the arguments accepted for the default
 #'   \code{plotfun="intervals"} can be found at
-#'   \code{\link[bayesplot]{mcmc_intervals}}.
+#'   \code{\link[bayesplot:MCMC-intervals]{mcmc_intervals}}.
 #'
 #' @return Either a ggplot object that can be further customized using the
 #'   \pkg{ggplot2} package, or an object created from multiple ggplot objects
@@ -53,7 +53,7 @@
 #'   \item The vignettes in the \pkg{bayesplot} package for many examples.
 #'   \item \code{\link[bayesplot]{MCMC-overview}} (\pkg{bayesplot}) for links to
 #'   the documentation for all the available plotting functions.
-#'   \item \code{\link[bayesplot]{color_scheme_set}} (\pkg{bayesplot}) to change
+#'   \item \code{\link[bayesplot:bayesplot-colors]{color_scheme_set}} (\pkg{bayesplot}) to change
 #'   the color scheme used for plotting.
 #'   \item \code{\link{pp_check}} for graphical posterior predictive checks.
 #'   \item \code{\link{plot_nonlinear}} for models with nonlinear smooth 
@@ -63,6 +63,7 @@
 #' @template reference-bayesvis
 #' 
 #' @examples
+#' if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
 #' \donttest{
 #' # Use rstanarm example model
 #' if (!exists("example_model")) example(example_model)
@@ -162,7 +163,7 @@
 #'
 #' # For graphical posterior predictive checks see
 #' # help("pp_check.stanreg")
-#'
+#' }
 #' @importFrom ggplot2 ggplot aes_string xlab %+replace% theme
 #'
 plot.stanreg <- function(x, plotfun = "intervals", pars = NULL,
@@ -314,7 +315,7 @@ validate_plotfun_for_opt_or_vb <- function(plotfun) {
 # pairs method ------------------------------------------------------------
 #' Pairs method for stanreg objects
 #' 
-#' Interface to \pkg{bayesplot}'s \code{\link[bayesplot]{mcmc_pairs}} function 
+#' Interface to \pkg{bayesplot}'s \code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}} function 
 #' for use with \pkg{rstanarm} models. Be careful not to specify too
 #' many parameters to include or the plot will be both hard to read and slow to
 #' render.
@@ -333,7 +334,7 @@ validate_plotfun_for_opt_or_vb <- function(plotfun) {
 #'   parameters it may be far too many to visualize on a small computer screen 
 #'   and also may require substantial computing time.
 #' @param condition Same as the \code{condition} argument to 
-#'   \code{\link[bayesplot]{mcmc_pairs}} except the \emph{default is different}
+#'   \code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}} except the \emph{default is different}
 #'   for \pkg{rstanarm} models. By default, the \code{mcmc_pairs} function in
 #'   the \pkg{bayesplot} package plots some of the Markov chains (half, in the
 #'   case of an even number of chains) in the panels above the diagonal and the
@@ -345,8 +346,9 @@ validate_plotfun_for_opt_or_vb <- function(plotfun) {
 #'   \code{accept_stat__} and the plots above the diagonal will contain 
 #'   realizations that are above the median \code{accept_stat__}. To change this
 #'   behavior see the documentation of the \code{condition} argument at 
-#'   \code{\link[bayesplot]{mcmc_pairs}}.
-#' @param ... Optional arguments passed to \code{\link[bayesplot]{mcmc_pairs}}. 
+#'   \code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}}.
+#' @param ... Optional arguments passed to 
+#'   \code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}}. 
 #'   The \code{np}, \code{lp}, and \code{max_treedepth} arguments to 
 #'   \code{mcmc_pairs} are handled automatically by \pkg{rstanarm} and do not 
 #'   need to be specified by the user in \code{...}. The arguments that can be 
@@ -354,10 +356,11 @@ validate_plotfun_for_opt_or_vb <- function(plotfun) {
 #'   \code{off_diag_fun}, \code{diag_args}, \code{off_diag_args},
 #'   and \code{np_style}. These arguments are
 #'   documented thoroughly on the help page for
-#'   \code{\link[bayesplot]{mcmc_pairs}}.
+#'   \code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}}.
 #' 
 #'   
 #' @examples
+#' if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
 #' \donttest{
 #' if (!exists("example_model")) example(example_model)
 #' 
@@ -414,7 +417,7 @@ validate_plotfun_for_opt_or_vb <- function(plotfun) {
 #' )
 #' 
 #' }
-#'
+#' }
 pairs.stanreg <-
   function(x,
            pars = NULL,